% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/authorize.R
\name{authorize}
\alias{authorize}
\title{Generate an OAuth 2.0 access token}
\usage{
authorize(.conn)
}
\arguments{
\item{.conn}{A \code{force_api} object.}
}
\description{
Uses the information stored in a \code{force_api} object to request an access
token from Salesforce.
}

