% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response-salesforce.R
\name{check_salesforce_response}
\alias{check_salesforce_response}
\title{HTTP error checking}
\usage{
check_salesforce_response(res)
}
\description{
Better handling for Salesforce error responses.
}

