% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect_to.R
\name{connect_to}
\alias{connect_to}
\title{Connect to a Salesforce instance}
\usage{
connect_to(url, ...)
}
\arguments{
\item{url}{A character vector of the instance URL}

\item{\dots}{Additional parameters required to authenticate the Force REST
  API, such as username, password, client_id, or client_secret}
}
\value{
A \code{force_api} R6 object
}
\description{
Start up a connection to a Salesforce instance.
}

