% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe a Salesforce object}
\usage{
describe(.conn, obj)
}
\arguments{
\item{.conn}{A \code{force_api} object.}

\item{obj}{A character vector, name of Salesforce object}
}
\description{
Lists all metadata for a Salesforce object. For example, \code{describe} be
used to retrieve object fields, URLs, and child relationships.
}

