% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/force_api.R
\name{force_api}
\alias{force_api}
\title{Force API object}
\usage{
force_api(base, version = NULL, username = NULL, password = NULL,
  client_id = NULL, client_secret = NULL)
}
\description{
This object holds the meta information for and represents a connection to the
}

