% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api-utils.R
\name{password}
\alias{password}
\title{Set password}
\usage{
password(.conn, psswd)
}
\arguments{
\item{.conn}{A \code{force_api} object}

\item{psswd}{A character vector, the password concatenated with the security
  token for \code{username}}
}
\description{
Set password
}

