% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Execute a SOQL query}
\usage{
query(.conn, qry, get_next = TRUE, verbose = TRUE)
}
\arguments{
\item{.conn}{A \code{force_api} object.}

\item{qry}{A character vector SOQL query, see SOQL section below for
details.}

\item{get_next}{A boolean value indicating that next records URLs should be
  followed}
}
\value{
A list of two values, \code{next_records_url} and \code{results}.
  NOTE, if \code{get_next} is \code{TRUE}, \code{next_records_url} will be
  \code{NA}.
}
\description{
Passes a SOQL query to the Force REST API. If the results of the query are
too large then a next records url is returned along with the partial results.
}

