% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api-utils.R
\name{use_version}
\alias{use_version}
\title{Specify API version}
\usage{
use_version(.conn, vrsn)
}
\arguments{
\item{.conn}{A \code{force_api} object}

\item{vrsn}{A character vector specifying the API version number or
  \code{"latest"}}
}
\value{
An updated \code{force_api} object
}
\description{
Adds a specifc version number to a \code{force_api} object or retrieves the
most recent API version.
}
\details{
The format for the version number requires a trailing decimal point
  and zero are appended, for example "23.0".

  If version is \code{"latest"} then the latest API version for the
  Salesforce instance specified by \code{.conn} is retrieved and assigned as
  the API version.
}

