% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api-utils.R
\name{username}
\alias{username}
\title{Specifiy Salesforce username}
\usage{
username(.conn, nm)
}
\arguments{
\item{.conn}{A \code{force_api} object}

\item{nm}{A character vector, the Salesforce username to connect with}
}
\value{
An updated \code{force_api} object
}
\description{
Add a username to a \code{force_api} object. This username will be used to
authenticate the REST API connection and generate a session id.
}

