% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/force_api.R
\docType{data}
\name{force_api}
\alias{force_api}
\title{Force.com API abstraction object}
\format{An R6Class object.}
\usage{
force_api
}
\description{
This object is an abstract representation of the Force.com REST API. The
object holds meta information necessary to use REST API resources and
provides helpful internal function. These functions are in turn used to
create S3-stylized functions which are exposed to the programmer.
}
\details{
Create a new \code{force_api} object with the \code{connect_to} function.
}
\section{Methods}{

\itemize{
 \item TODO
 \item TODO
}
}
\keyword{internal}

